requirements.txt
Django
djangorestframework
django-crispy-forms
dj-database-url
psycopg2-binary
gunicorn
whitenoise
